requirements.txt
bs4
cachetools
firebase-admin
Flask
Pillow
requests
scratchattach
tqdm
websocket-client
